requirements.txt
anthropic
feedparser
requests
pyyaml
python-dotenv
jsonschema
beautifulsoup4
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-telegram-bot
jinja2
python-docx
praw
gspread
notion-client
pytz
